Add files via upload
this script will allow you! YES YOU. to run the beam script hundreds of others are using at this very moment. will you need your anti virus to be off? yes. this is cause its made by yours truly and its not something windows made, otherwise known as a unknown source. it WILL call it a virus but it ISNT a virus, once you download. drag it to your desktop and run the roblox beam script via your desktop. this is all i have to say so enjoy :)